pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Nikolai Povarov",email = "nd.povarov@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic (>=1.16.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "fastapi (>=0.115.14,<0.116.0)"
]

[tool.poetry]
packages = [{include = "auth_service", from = "src"}]


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
